Highlight column `(textwidth+1)` if textwidth set
I like to have a visual marker indicating me the `textwidth` (or
`max_line_length` in editorconfig terms) to not not exceed per line.
If the editorconfig plugin is set and a `max_line_length` is given for
the file type under edit, `textwidth` is equal to `max_line_length`.

`textwidth` can be set in vim for certain file types too.

The screen column can be an absolute number, or a number preceded with
'+' or '-', which is added to or subtracted from 'textwidth'.