Add mapping to project wide rename current symbol
It fires a `:CocSearch` with the current word. If you need more context
to be shown below the match use `-A numberOfLinesBelow` to show match
and the numberOfLines below it (-B for Above). This is all due to the
fact it uses `rg` under the hood.

Check https://youtu.be/q7gr6s8skt0?t=159 for refresher on how the
CocSearch works.