Fix ESC key delay by decreasing escape-time
When keyboards do not have Meta key (e.g Alt), this function can be
covered by using Esc as the Modifier, meaning Esc+a would be equivalent
to Alt+a.
To make this possible an escape-time delay is set by Tmux (defaulted to
500ms). This way if a second key stroke occures before the delay
expires, the Meta key sequence is issued.

Having a too big delay had Vim to ignore Esc key if I was too quick
typing something else.
For instance:

- Go to insert mode
- Add random text
- Press Escape to exit insert mode
- <Leader>w to Save
- You are still stuck in insert mode :(

Lowering the escape-time to 20ms fixed this issue!